backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
cryptography==41.0.7
requests==2.31.0
pyOpenSSL==23.3.0
click==8.1.7
Werkzeug==2.3.7
validators==0.22.0
python-dateutil==2.8.2
dnspython==2.4.2
gunicorn==21.2.0

